requirements.txt
celery
redis
tiktoken
tree-sitter==0.20.1
tree-sitter-languages==1.7.0
langchain
langchain-openai
qdrant-client
gitpython
networkx
prometheus-client
openai
GitPython
requests